fix: complete UE5.7 + macOS ARM link fixes (LibCarla, linker, foliage)
LibCarla / Boost 1.87+ / macOS ARM fixes:
- Waypoint.cpp: use std::size_t seed for boost::hash_combine (uint64_t ≠ size_t on ARM)
- listener.cpp: io_context::reset() → restart() (Boost 1.66+)
- secondary.cpp, router.cpp: from_string() → make_address() (Boost 1.66+)
- primaryCommands.cpp, router.cpp: Buffer(ptr, size_t) ambiguous → cast to uint64_t
- Buffer.h, MsgPack.h, MsgPackAdaptors.h, ThreadPool.h, rpc/Server.h,
  streaming/EndPoint.h, streaming/detail/tcp/Server.h: UE5/C++20 compat

Unreal linker fixes:
- Carla.Build.cs: add DeveloperToolSettings for UProjectPackagingSettings (moved in UE5)
- Carla.Build.cs: guard CarlaDependencies libs with File.Exists(), libsqlite3 .so/.a fallback,
  PyTorch/CUDA Linux-only, OsmToODR libs conditional
- CarlaTools.Build.cs: remove FoliageEdit/Private include path (hidden symbol)
- MapGeneratorWidget.cpp: replace FEdModeFoliage::AddInstances (non-exported hidden symbol)
  with public API: FoliageTrace + FPotentialInstance::PlaceInstance + FFoliageInfo::AddInstances
- EditorCamera.cpp: Cast<FEditorViewportClient> → static_cast (non-UObject)
- ProceduralBuildingUtilities.cpp: FSavePackageArgs, missing includes, private member fixes
- MapGeneratorWidget.cpp: VectorRegister ctor → FTransform, FSavePackageArgs, ALandscapeProxy
  new Import param, UActorFactory::CreateActor → FActorSpawnParameters
- MapPreviewUserWidget.cpp: buffer_cast → data().data(), Texture->GetResource()
- WorldObserver.cpp: buffer.reset(uint64_t) cast
- LineBatchComponent_CARLA: FVector3f/FVector2f for DynamicMeshBuilder (UE5)
- CarlaUE4.h: forward-declare Algo::Sort to break circular include (UE5.7 bug)

Other:
- MDL/TextureExporterDDS → MDLTextureExporterDDS (rename with MDL prefix)
- build_carla_deps_macos.sh: script to build rpclib/libcarla_server/osm2odr on macOS ARM
- CarlaUE4.uproject: engine 5.7, AirSim disabled, plugin list updated
- VehicleAckermannControl.h: enable/disable UE4 macros wrapper

Result: Result: Succeeded (full clean build on macOS ARM UE5.7)

+++ LibCarla/source/carla/multigpu/primaryCommands.cpp
@@ -36,7 +36,7 @@ void PrimaryCommands::SendFrameData(carla::Buffer buffer) {
 
 // broadcast to all secondary servers the map to load
 void PrimaryCommands::SendLoadMap(std::string map) {
-  carla::Buffer buf((unsigned char *) map.c_str(), (size_t) map.size() + 1);
+  carla::Buffer buf((unsigned char *) map.c_str(), (uint64_t) map.size() + 1);
   _router->Write(MultiGPUCommand::LOAD_MAP, std::move(buf));
 }
 
@@ -44,7 +44,7 @@ void PrimaryCommands::SendLoadMap(std::string map) {
 token_type PrimaryCommands::SendGetToken(stream_id stream_id) {
   log_info("asking for a token");
   carla::Buffer buf((carla::Buffer::value_type *) &stream_id,
-                    (size_t) sizeof(stream_id));
+                    (uint64_t) sizeof(stream_id));
   auto fut = _router->WriteToNext(MultiGPUCommand::GET_TOKEN, std::move(buf));
 
   auto response = fut.get();
@@ -56,7 +56,7 @@ token_type PrimaryCommands::SendGetToken(stream_id stream_id) {
 // send to know if a connection is alive
 void PrimaryCommands::SendIsAlive() {
   std::string msg("Are you alive?");
-  carla::Buffer buf((unsigned char *) msg.c_str(), (size_t) msg.size());
+  carla::Buffer buf((unsigned char *) msg.c_str(), (uint64_t) msg.size());
   log_info("sending is alive command");
   auto fut = _router->WriteToNext(MultiGPUCommand::YOU_ALIVE, std::move(buf));
   auto response = fut.get();
@@ -68,7 +68,7 @@ void PrimaryCommands::SendEnableForROS(carla::streaming::detail::stream_actor_id
   auto it = _servers.find(stream_actor_id.stream_id);
   if (it != _servers.end()) {
     carla::Buffer buf((carla::Buffer::value_type *) &stream_actor_id,
-                      (size_t) (sizeof(stream_actor_id)));
+                      (uint64_t) (sizeof(stream_actor_id)));
     auto fut = _router->WriteToOne(it->second, MultiGPUCommand::ENABLE_ROS, std::move(buf));
 
     auto response = fut.get();
@@ -83,7 +83,7 @@ void PrimaryCommands::SendDisableForROS(carla::streaming::detail::stream_actor_i
   auto it = _servers.find(stream_actor_id.stream_id);
   if (it != _servers.end()) {
     carla::Buffer buf((carla::Buffer::value_type *) &stream_actor_id,
-                      (size_t) (sizeof(stream_actor_id)));
+                      (uint64_t) (sizeof(stream_actor_id)));
     auto fut = _router->WriteToOne(it->second, MultiGPUCommand::DISABLE_ROS, std::move(buf));
 
     auto response = fut.get();
@@ -98,7 +98,7 @@ bool PrimaryCommands::SendIsEnabledForROS(carla::streaming::detail::stream_actor
   auto it = _servers.find(stream_actor_id.stream_id);
   if (it != _servers.end()) {
     carla::Buffer buf((carla::Buffer::value_type *) &stream_actor_id,
-                      (size_t) (sizeof(stream_actor_id)));
+                      (uint64_t) (sizeof(stream_actor_id)));
     auto fut = _router->WriteToOne(it->second, MultiGPUCommand::IS_ENABLED_ROS, std::move(buf));
 
     auto response = fut.get();
